Supress a warning in VS-based builds.
Thanks @xfury for pointing out.

+++ libmpq/libmpq/mpq.c
@@ -867,7 +867,7 @@ int32_t libmpq__block_read(mpq_archive_s *mpq_archive, uint32_t file_number, uin
 	uint32_t imploded   = 0;
 	int32_t tb          = 0;
 	libmpq__off_t block_offset  = 0;
-	off_t in_size       = 0;
+	size_t in_size      = 0;
 	libmpq__off_t unpacked_size = 0;
 
 	/* check if given file number is not out of range. */
